requirements.txt
altair==5.2.0
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli @ file:///C:/ci_311/brotli-split_1676435766766/work
cachetools==5.3.2
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.3
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fsspec==2023.10.0
gdown==5.0.0
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.3
idna @ file:///C:/ci_311/idna_1676424932545/work
importlib-metadata==7.0.1
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
Jinja2 @ file:///C:/ci_311/jinja2_1676424968965/work
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.1.0
jupyter_core==5.3.0
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci_311/markupsafe_1676424152318/work
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
nest-asyncio==1.5.6
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy==1.26.1
packaging==23.0
pandas==2.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==4.25.2
psutil==5.9.4
pure-eval==0.2.2
pyarrow==15.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.1b0
Pygments==2.14.0
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==305
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
pyzmq==25.0.2
referencing==0.32.1
regex==2023.10.3
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.2
streamlit==1.30.0
sympy @ file:///C:/ci_311_rebuilds/sympy_1679009400182/work
tenacity==8.2.3
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.37.0
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zipp==3.17.0
